+++ workspace/apps/server/src/contexts/session.ts
@@ -15,11 +15,14 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-console.log(import.meta.filename);
+import type { CreateFastifyContextOptions } from "@trpc/server/adapters/fastify";
 
-async function prod() {
+/**
+ * @see {@link https://trpc.io/docs/server/adapters/fastify#create-the-context Create the context}
+ */
+export function createSessionContext(options: CreateFastifyContextOptions) {
+    // TODO: 从 Cookies 中获取 JWT 并解析
+    return options;
 }
 
-if (process.argv.includes(import.meta.filename)) {
-    prod();
-}
+export type TSessionContext = Awaited<ReturnType<typeof createSessionContext>>;